+++ src/RcppExports.cpp
@@ -13,6 +13,18 @@ Rcpp::Rostream<true>&  Rcpp::Rcout = Rcpp::Rcpp_cout_get();
 Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 #endif
 
+// g_index_
+int g_index_(const Rcpp::NumericVector& citations, bool is_sorted);
+RcppExport SEXP _eddington_g_index_(SEXP citationsSEXP, SEXP is_sortedSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< const Rcpp::NumericVector& >::type citations(citationsSEXP);
+    Rcpp::traits::input_parameter< bool >::type is_sorted(is_sortedSEXP);
+    rcpp_result_gen = Rcpp::wrap(g_index_(citations, is_sorted));
+    return rcpp_result_gen;
+END_RCPP
+}
 // E_num
 int E_num(const Rcpp::IntegerVector& rides);
 static SEXP _eddington_E_num_try(SEXP ridesSEXP) {
@@ -42,7 +54,7 @@ RcppExport SEXP _eddington_E_num(SEXP ridesSEXP) {
     if (rcpp_isError_gen) {
         SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
         UNPROTECT(1);
-        Rf_error("%s", CHAR(rcpp_msgSEXP_gen));
+        (Rf_error)("%s", CHAR(rcpp_msgSEXP_gen));
     }
     UNPROTECT(1);
     return rcpp_result_gen;
@@ -76,7 +88,7 @@ RcppExport SEXP _eddington_E_cum(SEXP ridesSEXP) {
     if (rcpp_isError_gen) {
         SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
         UNPROTECT(1);
-        Rf_error("%s", CHAR(rcpp_msgSEXP_gen));
+        (Rf_error)("%s", CHAR(rcpp_msgSEXP_gen));
     }
     UNPROTECT(1);
     return rcpp_result_gen;
@@ -110,7 +122,7 @@ RcppExport SEXP _eddington_E_next(SEXP ridesSEXP) {
     if (rcpp_isError_gen) {
         SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
         UNPROTECT(1);
-        Rf_error("%s", CHAR(rcpp_msgSEXP_gen));
+        (Rf_error)("%s", CHAR(rcpp_msgSEXP_gen));
     }
     UNPROTECT(1);
     return rcpp_result_gen;
@@ -154,6 +166,7 @@ RcppExport SEXP _eddington_RcppExport_registerCCallable() {
 RcppExport SEXP _rcpp_module_boot_eddington_mod();
 
 static const R_CallMethodDef CallEntries[] = {
+    {"_eddington_g_index_", (DL_FUNC) &_eddington_g_index_, 2},
     {"_eddington_E_num", (DL_FUNC) &_eddington_E_num, 1},
     {"_eddington_E_cum", (DL_FUNC) &_eddington_E_cum, 1},
     {"_eddington_E_next", (DL_FUNC) &_eddington_E_next, 1},